`name` members attached to Argument members.
In register() function, all the named Arguments in self.Registry had
the that string attached as `name` to it.

Also, the character check on the `short` form name of the Arguments
originalling in the Argument.__init__() method no performed at the
end of register() (so we know what the `name` is!).